% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cepstral.R
\name{generate_sig}
\alias{generate_sig}
\title{Generate Exponential Correlation Covariance Matrix}
\usage{
generate_sig(n, rho)
}
\arguments{
\item{n}{Dimension of the covariance matrix.}

\item{rho}{Correlation decay parameter.}
}
\value{
An n × n positive definite covariance matrix.
}
\description{
Creates an n × n covariance matrix with entries \eqn{\rho^{|i - j|}}.
}
\examples{
S <- generate_sig(5, 0.5)


}
